+++ docs/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Changelog for v2.2.0
 + Remove the mechanism to manage non-standardized prj file
 + DBFReader uses VarChar instead of Char for String values
++ Change move some log info to debug mode