+++ docs/CHANGELOG.md
@@ -2,3 +2,4 @@
 + Remove the mechanism to manage non-standardized prj file
 + DBFReader uses VarChar instead of Char for String values
 + Change move some log info to debug mode
++ Implements flatgeobuf writer and reader